requirements.txt
pybase64
pyyaml
schedule
requests
paho-mqtt